Decompose GitHubServiceTests into focused test suites
Split the 1800-line test file into 8 smaller, domain-specific files:

- BuildStatusPresentationTests (28 lines)
- NonBlockingCheckSummaryTests (117 lines)
- GitHubServiceHostCacheTests (30 lines)
- GitHubServiceFetchErrorTests (62 lines)
- GitHubServiceBatchQueryTests (76 lines)
- GitHubServiceBatchResponseParsingTests (215 lines)
- GitHubServiceFetchOtherPRTests (232 lines)
- GitHubServiceTests (core batch integration, 1009 lines)

Extract MockShellExecutor into its own file (internal access)
so it can be shared across test suites.
